--- sites/jeromefitzgerald.com/src/app/_v16/types.ts
@@ -1,19 +1,0 @@
-import type { StaticImageData } from "next/image";
-
-type Image = {
-  alt?: string;
-  blurDataURL?: string;
-  className?: string;
-  height?: number;
-  order?: number;
-  quality?: number;
-  sizes?: string;
-  src: StaticImageData | string;
-  width?: number;
-};
-
-const blurDataURL =
-  "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAoAAAAICAIAAABPmPnhAAAACXBIWXMAAAsTAAALEwEAmpwYAAAA1UlEQVR4nFWOwUrDQABEZ3az2e3awC6VpI3NHxS00KOXtpdKSiSo6bkKBW+evfgNlebuRezJP9EP6K94qJiK4LsNMzAP/I8gAcRJ//GpvsgXYJMPkPC+U17drZ+3m83reJz/FT+jLM2qavn5sasuyyTOJtPqt9bGAGjZdjG/qdd12nHvL2/Xi3sczoy1JI+i6HQwnA3Pi3m+/9qvbh+gtQYgVUDSO3fcPfFJ2ku73ruzwQham1ApIYWQsmWNixPrHNjYkBAUKgxFIAmqMLBRW0jJRpbkNwFLHj/O9IP8AAAAAElFTkSuQmCC";
-
-export type { Image };
-export { blurDataURL };
