+++ sites/jeromefitzgerald.com/src/components/v16/current.home.tsx
